+++ src/main/java/net/kaczmarzyk/spring/data/jpa/domain/DateBefore.java
@@ -43,7 +43,7 @@ public class DateBefore<T> extends DateSpecification<T> {
             throw new IllegalArgumentException("expected a single http-param, but was: " + args);
         }
         String dateStr = args[0];
-        this.date = format.parse(dateStr);
+        this.date = converter.convertToDate(dateStr);
     }
 
     @Override