+++ src/test/java/net/kaczmarzyk/spring/data/jpa/CustomerBuilder.java
@@ -48,9 +48,18 @@ public class CustomerBuilder {
         cal.set(Calendar.YEAR, year);
         cal.set(Calendar.MONTH, month - 1);
         cal.set(Calendar.DAY_OF_MONTH, day);
+        cal.set(Calendar.HOUR_OF_DAY, 0);
+        cal.set(Calendar.MINUTE, 0);
+        cal.set(Calendar.SECOND, 0);
+        cal.set(Calendar.MILLISECOND, 0);
         customer.setRegistrationDate(cal.getTime());
         return this;
     }
+
+    public CustomerBuilder weight(int weight) {
+		customer.setWeight(weight);
+		return this;
+	}
     
     public Customer build(EntityManager em) {
         em.persist(customer);