+++ CHANGELOG.md
@@ -2,3 +2,4 @@ v0.5.0
 ======
 
 * introduced `DateAfter` specification
+* introduced `Equal` that supports exact match for numbers, strings, dates and enums